Consolidate deployable entity tests
