+++ client/verta/tests/deployable_entity/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Inherited/implemented functionality of _DeployableEntity subclasses."""
